+++ CHANGELOG.md
@@ -10,16 +10,17 @@ Change Log
 5. Added bnd build tool and `new.sh` script;
 6. Added IntelliJ IDEA support;
 7. Added Maven bnd repo support;
-8. Added Maven library support; (IntelliJ)
+8. Added Maven library support; (Gradle)
 9. Added hot deploy setup;
 10. Added `bundle.sh` script;
 11. Added IntelliJ official .gitignore piece;
 12. Added headless sub project compile; (Gradle)
+13. Added headless sub project lib sync; (Gradle)
+14. Added Scala support to sub project; (Gradle)
 
 working on ...
 
 * gradle task to call bundle.sh; (override jar task)
-* gradle sub project support +Scala;
 * docker containers;
 * session sub project; (account/user, role/permission ==> jwttoken, session)
 * db sub project; (K-V, JSON/Doc, SQL, BigTable, Graph)